move correlation actions cell to separate file
