+++ frontend/src/scenes/insights/views/Funnels/CorrelationActionsCell.tsx
@@ -1,0 +1,58 @@
+import { useState } from 'react'
+import { Row } from 'antd'
+import { useActions, useValues } from 'kea'
+import { EllipsisOutlined } from '@ant-design/icons'
+import { funnelLogic } from 'scenes/funnels/funnelLogic'
+import { FunnelCorrelation, FunnelCorrelationResultsType } from '~/types'
+import { insightLogic } from 'scenes/insights/insightLogic'
+import { LemonButton } from 'lib/lemon-ui/LemonButton'
+import { Popover } from 'lib/lemon-ui/Popover/Popover'
+
+export const CorrelationActionsCell = ({ record }: { record: FunnelCorrelation }): JSX.Element => {
+    const { insightProps } = useValues(insightLogic)
+    const logic = funnelLogic(insightProps)
+    const { excludeEventPropertyFromProject, excludeEventFromProject, setFunnelCorrelationDetails } = useActions(logic)
+    const { isEventPropertyExcluded, isEventExcluded } = useValues(logic)
+    const components = record.event.event.split('::')
+    const [popoverOpen, setPopoverOpen] = useState(false)
+
+    return (
+        <Row style={{ justifyContent: 'flex-end' }}>
+            <Popover
+                visible={popoverOpen}
+                actionable
+                onClickOutside={() => setPopoverOpen(false)}
+                overlay={
+                    <>
+                        {record.result_type === FunnelCorrelationResultsType.Events && (
+                            <LemonButton onClick={() => setFunnelCorrelationDetails(record)} fullWidth status="stealth">
+                                View correlation details
+                            </LemonButton>
+                        )}
+                        <LemonButton
+                            disabled={
+                                record.result_type === FunnelCorrelationResultsType.EventWithProperties
+                                    ? isEventPropertyExcluded(components[1])
+                                    : isEventExcluded(components[0])
+                            }
+                            onClick={() =>
+                                record.result_type === FunnelCorrelationResultsType.EventWithProperties
+                                    ? excludeEventPropertyFromProject(components[0], components[1])
+                                    : excludeEventFromProject(components[0])
+                            }
+                            fullWidth
+                            title="Remove this event from any correlation analysis report in this project."
+                            status="stealth"
+                        >
+                            Exclude event from project
+                        </LemonButton>
+                    </>
+                }
+            >
+                <LemonButton status="stealth" onClick={() => setPopoverOpen(!popoverOpen)}>
+                    <EllipsisOutlined className="insight-dropdown-actions" />
+                </LemonButton>
+            </Popover>
+        </Row>
+    )
+}
